router: test around inherent race condition
Though the server has closed this conn, the http.Transport may not have
removed it from its conn pool before it attempts to send a request
through to it. That's because the persistent conn loop may not have
ACK'd the TCP FIN from the server by the time it handles our request.

Work around that with a sleep for now. Hopefully it will be fixed by
https://golang.org/issue/4677

Signed-off-by: Blake Gentry <blakesgentry@gmail.com>