router: use proxy package in httpService
Add a proxy object to httpService. The proxy object is either a dynamic
or sticky router.Proxy.

Fix the http tests to work with the new router/proxy package. Most of
the changes are related to values that have been scoped to the proxy
package and are no longer accessable from the router package.

Signed-off-by: Ben Burkert <ben@benburkert.com>